Translate attempt tables to hyper3 support functions

description here

attemptstable2supp3(a, decreasing, give.supp=TRUE, dnf.last=TRUE)

Arguments

a

Data frame, see details

decreasing

Boolean, with TRUE meaning that the highest score wins [e.g. javelin distances] and FALSE meaning that the lowest score wins [e.g. times for a race]

give.supp

Boolean, return the support function or the order statistic

dnf.last

Boolean, should NA entries count as coming last (TRUE) or be ignored (FALSE)

Details

Function attemptstable2supp3() is intended for use on attempts tables like javelin.

These objects can be generated by running script inst/javelin.Rmd, which includes some further discussion and technical documentation, and creates file javelin.rda which resides in the data/ directory.

Value

Returns a hyper3 object

Author

Robin K. S. Hankin

See also

Examples

jj <- javelin_table[1:3,]
jj
#>          throw1 throw2 throw3 throw4 throw5 throw6
#> Chopra    87.03  87.58  76.79      X      X  84.24
#> Vadlejch  83.98      X      X  82.86  86.67      X
#> Vesely    79.73  80.30  85.44      X  84.98      X
attemptstable2supp3(jj)
#> log( (Chopra=1)^4 * (Chopra=3, Vadlejch=3, Vesely=2)^-1 * (Chopra=3,
#> Vadlejch=3, Vesely=3)^-1 * (Chopra=3, Vadlejch=3, Vesely=4)^-1 *
#> (Chopra=3, Vadlejch=4, Vesely=4)^-1 * (Chopra=3, Vadlejch=5,
#> Vesely=4)^-1 * (Chopra=4, Vadlejch=5, Vesely=4)^-1 * (Chopra=4,
#> Vadlejch=5, Vesely=5)^-1 * (Chopra=4, Vadlejch=5, Vesely=6)^-1 *
#> (Chopra=4, Vadlejch=6, Vesely=6)^-1 * (Chopra=5, Vadlejch=6,
#> Vesely=6)^-1 * (Chopra=6, Vadlejch=6, Vesely=6)^-1 * (Vadlejch=1)^3 *
#> (Vesely=1)^4)