Chess playing dataset

A tally of wins and losses for games between three chess players: Topalov, Anand, Karpov.

data(chess)

Details

(there are three chess datasets in the package, documented at interzonal.Rd [the 1963 World championship], kka.Rd [Karpov-Kasparov-Anand dataset], and chess.Rd [rock-paper-scissors using Topalov-Anand-Karpov])

This is a very simple dataset that can be used for illustration of hyper2 idiom.

The players are:

Grandmaster Veselin Topalov. FIDE world champion 2005-2006; peak rating 2813

Grandmaster Viswanathan Anand. FIDE world champion 2000-2002, 2008; peak rating 2799

Grandmaster Anatoly Karpov. FIDE world champion 1993-1999; peak rating 2780

Observe that Topalov beats Anand, Anand beats Karpov, and Karpov beats Topalov (where “beats” means “wins more games than”).

The games thus resemble a noisy version of “rock paper scissors”.

The likelihood function does not record who played white; see karpov_kasparov_anand for such a dataset.

These objects can be generated by running script inst/rock_paper_scissors.Rmd, which includes some further discussion and technical documentation and creates file chess.rda which resides in the data/ directory.

File inst/ternaryplot_hyper2.Rmd gives an example showing the chess likelihood function that uses Ternary::ternaryPlot().

References

https://en.chessbase.com/

See also

Examples

data(chess)
maxp(chess)
#>   Topalov     Anand    Karpov 
#> 0.4036108 0.3405168 0.2558723 

 mgf(chess,c(Anand=2),tol = 0.1)  # tolerance for speed
#> [1] 1