Consistency check for hyper2 objects

Given a hyper2 object, calculate the maximum likelihood point in two ways and plot one against the other to check for consistency.

consistency(H, plot=TRUE, ...)

Arguments

H

A hyper2 object

plot

If TRUE (default), plot a comparison and return a matrix invisibly, and if FALSE return the matrix. Modelled on argument plot of hist

...

Further arguments, passed to points()

Details

Given a hyper2 object, calculate the maximum likelihood estimate of the players' strengths using maxp(); then reverse the pnames attribute and calculate the players' strengths again. These two estimates should be identical but small differences highlight numerical problems. Typically, the differences are small if there are fewer than about 25 players.

Reversing the pnames() is cosmetic in theory but is a non-trivial operation: for example, it changes the identity of the fillup from the last player to the first.

Value

Returns a named three-row matrix with first row being the direct evaluate, second row being the reverse of the reversed evaluate, and the third being the difference

Author

Robin K. S. Hankin

See also

Examples

# consistency(icons)

x <- icons
y <- icons
pnames(y) <- rev(pnames(y))
gradient(x,indep(equalp(x)))
#> [1] 132.0  82.5 118.5  69.0  30.0
gradient(y,indep(equalp(y)))
#> [1] -132.0 -102.0  -63.0  -13.5  -49.5