Wrappers to c calls

Various low-level wrappers to C functions, courtesy of Rcpp

overwrite(L1, powers1, L2, powers2)
accessor(L,powers,Lwanted)
assigner(L,p,L2,value)
addL(L1,p1,L2,p2)
identityL(L,p)
evaluate(L, powers, probs, pnames)
differentiate(L, powers, probs, pnames, n)
differentiate_n(L, powers, probs, pnames, n)

Arguments

L,L1,L2,Lwanted

Lists with character vector elements, used to specify the brackets of the hyperdirichlet distribution

p,p1,p2,powers,powers1,powers2

A numeric vector specifying the powers to which the brackets are raised

value

RHS in assignment, a numeric vector

probs

Vector of probabilities for evaluation of log-likelihood

pnames

Character vector of names

n

Integer specifying component to differentiate with respect to

Details

These functions are not really intended for the end-user, as out-of-scope calls may cause crashes.

Value

These functions return a named List

Author

Robin K. S. Hankin