Dataset on communication breakdown in handover between physicians

Object handover is a likelihood function corresponding to a dataset arising from 69 medical malpractice claims and concerns handover (or hand-off) between physicians. This dataset was analysed by Lin et al. (2009), and further analysed by Altham and Hankin (2010). The computational methods are presented in the (unmaintained) hyperdirichlet and aylmer packages and a further discussion is given in the “integration” vignette of the hyper2 package. The original dataset is handover_table, a three-by-three matrix of counts.

data(handover)

Details

These objects can be generated by running script inst/handover.Rmd, which includes some further discussion and technical documentation, and creates file handover.rda which resides in the data/ directory.

References

Y. Lin and S. Lipsitz and D. Sinha and A. A. Gawande and S. E. Regenbogen and C. C. Greenberg, 2009. “Using Bayesian \(p\)-values in a \(2\times 2\) table of matched pairs with incompletely classified data”. Journal of the Royal Statistical Society, Series C, 58:2

P. M. E. Altham and R. K. S. Hankin, 2010. “Using recently developed software on a \(2\times 2\) table of matched pairs with incompletely classified data”. Journal of the Royal Statistical Society, series C, 59(2): 377-379

R. K. S. Hankin 2010. “A generalization of the Dirichlet distribution”. Journal of Statistical software, 33:11

L. J. West and R. K. S. Hankin 2008. “Exact tests for two-way contingency tables with structural zeros”. Journal of Statistical software, 28:11

Examples

data(handover)
maxp(handover)
#>        t00        t01        t10        t11 
#> 0.41954894 0.11127554 0.01798719 0.45118833