Karate dataset

Dataset from the 2018 World Karate Championships, men's 67kg. It is an example of a dataset with too many degrees of freedom to be analysed easily by the package.

data(karate)

Details

Object karate_table is a dataframe of results showing results from the 2018 World Karate Championships, men's 67kg; karate is the associated likelihood function. There are two maximum likelihood estimates given; karate_maxp, the evaluate as returned by maxp(), and karate_maxp, returned by zermelo() [the value given by maxp() itself is less likely].

These objects can be generated by running script inst/karate.Rmd, which includes some further discussion and technical documentation and creates file karate.rda which resides in the data/ directory.

References

https://en.wikipedia.org/wiki/2018_World_Karate_Championships

Note

Table karate_table misses uninformative matches, that is, competitions with 0-0 results.

See also

Examples

summary(karate)
#> A hyper2 object of size 82.
#> pnames:  akondzo alhadharim alkhathami almasatfa avila balcius bodrovs bouakel catro cenaj cheung dacosta daraghma delgado derafshipour dimitrov diouf elsawy ferreras figueira fillies galvez gidakos gill gordillo gorenc gudmundsson gutierrez hasanov hodzic hollowood homola jihwan joksic karkkalainen kavanagh khani krautsou kuhn limam lindelauf loannides maresca matthiasen mendez mertel mora muratov murtazaliev narkiniemi ngamphuengphit nievas noriega nyoni oneil pak pataridze petkov pokorny poshen povrzenic rodionov rodriguiez rolle rosiello sabiecki sawadogo shafei sharma shinohara sijercic suleimani tadissi tawfik teodorescu thomas toli uygur uzakov velozo zakaria zhiwei 
#> Number of brackets: 147 
#> Sum of powers: 0 
#> 
#> Table of bracket lengths:
#>  1  2 
#> 61 86 
#> 
#> Table of powers:
#> -13 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   1   2   3   4   5   6   7   8 
#>   2   1   1   6   8   1   4  11  17  10   8  17   8  11   9   4   6   2   4   2 
#>   9  10  12  16  17  18  19  20  31 
#>   3   3   1   2   2   1   1   1   1