Keep or discard players

Flawed functionality to keep or discard subsets of the players in a hyper2 object or order table.

discard_flawed2(x, unwanted,...)
keep_flawed(H, wanted)
discard_flawed(H, unwanted)

Arguments

H

A hyper2 object

x

An order table

wanted, unwanted

Players to keep or discard. May be character or integer or logical

...

Further arguments passed to wikitable_to_ranktable(), notably points

Details

Do not use these functions. They are here as object lessons in poor thinking. To work with a subset of competitors, see the example at as.ordertable.

Functions keep_flawed2() and discard_flawed2() take an order table and keep or discard specified rows, returning a reduced order table. This is not a trivial operation.

Functions keep_flawed() and discard_flawed() will either keep or discard players specified in the second argument. It is not clear to me that these functions have any reasonable probabilistic interpretation and file inst/retain.Rmd gives a discussion.

Given a wikitable or ordertable, it is possible to create a likelihood function based on a subset of rows using the incomplete=TRUE argument; see the example at ?ordertable2supp. But this method is flawed too because it treats non-finishers as if they finished in the order of their rows.

Function as.ordertable() is the correct way to consider a subset of players in a wikitable.

Author

Robin K. S. Hankin

See also

Examples

maxp(icons)
#>         NB          L         PB        THC         OA       WAIS 
#> 0.25230411 0.17364433 0.22458188 0.17011281 0.11068604 0.06867083 
discard_flawed(icons, c("OA", "WAIS"))
#> Warning: powers have nonzero sum
#> log(L^24 * (L + NB)^-9 * (L + NB + THC)^-35 * (L + PB)^-18 * (L + PB +
#> THC)^-27 * NB^32 * (NB + PB)^-8 * (NB + PB + THC)^-36 * PB^30 * THC^24)

if (FALSE) # (takes too long)
data("skating")
maxp(skating)[1:4]      # numbers work, keep the first four skaters
#>    babiakova   butyrskaya        cohen      fontana 
#> 3.528153e-06 1.071112e-02 7.494685e-02 8.014398e-04 
maxp(keep_flawed(skating, pnames(skating)[1:4])) # differs!
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in loglik_single(p, H, log = log) : sum(powers(H)) == 0 is not TRUE
#> known R bug (bugzilla ID 17703; wmmin not finite).  Kludge:  maxp_single() will try  a slightly different start point
#> Error in ms(H, startp = startp, give = TRUE, fcm = fcm, fcv = fcv, SMALL = SMALL,     ...): known R bug (bugzilla ID 17703).  Try increasing maxtry.
 # \dontrun{}