Sushi dataset

Dataset from the preflib website, discussing sushi preferences.

data(sushi)

Details

Sushi preference dataset.

Object sushi_table is a database of sushi preferences; each row is a judge and the rows are rank vectors of the different sushi types.

Object sushi is a likelihood function at the estimated value of lambda

Object sushi_maxp is the maximum likelihood estimate for the Bradley-Terry strengths of the sushis, at the estimate for lambda.

sushi_eq_classes is a named vector giving the equivalences classes of sushi. Class 1 is tuna, class 2 is everything else.

These objects can be generated by running script inst/sushi.Rmd, which includes some further discussion and technical documentation and creates file sushi.rda which resides in the data/ directory.

Examples

sushi
#> log( (anago=1)^36 * (anago=1, ebi=1, ika=1, kappa=1, maguro=1, sake=1,
#> tamago=1, tekka=1, toro=1)^-2 * (anago=1, ebi=1, ika=1, kappa=1,
#> maguro=1, sake=1, tamago=1, tekka=1, toro=1, uni=1)^-50 * (anago=1,
#> ebi=1, ika=1, kappa=1, maguro=1, sake=1, tekka=1, toro=1)^-2 *
#> (anago=1, ebi=1, ika=1, kappa=1, maguro=1, sake=1, tekka=1, toro=1,
#> uni=1)^-21 * (anago=1, ebi=1, ika=1, kappa=1, maguro=1, tamago=1,
#> tekka=1, toro=1, uni=1)^-1 * (anago=1, ebi=1, ika=1, kappa=1, sake=1,
#> tamago=1, tekka=1, toro=1, uni=1)^-3 * (anago=1, ebi=1, ika=1, kappa=1,
#> sake=1, tekka=1, toro=1)^-1 * (anago=1, ebi=1, ika=1, kappa=1, sake=1,
#> tekka=1, toro=1, uni=1)^-4 * (anago=1, ebi=1, ika=1, kappa=1, tamago=1,
#> tekka=1, toro=1)^-1 * (anago=1, ebi=1, ika=1, kappa=1, tamago=1,
#> tekka=1, toro=1, uni=1)^-1 * (anago=1, ebi=1, ika=1, maguro=1, sake=1,
#> tamago=1, tekka=1, toro=1, uni=1)^-2 * (anago=1, ebi=1, ika=1,
#> maguro=1, sake=1, tekka=1, toro=1)^-1 * (anago=1, ebi=1, ika=1,
#> maguro=1, sake=1, tekka=1, toro=1, uni=1)^-2 * (anago=1, ebi=1,
#> kappa=1, maguro=1, sake=1, tamago=1, tekka=1, toro=1)^-2 * (anago=1,
#> ebi=1, kappa=1, maguro=1, sake=1, tamago=1, tekka=1, toro=1, uni=1)^-12
#> * (anago=1, ebi=1, kappa=1, maguro=1, sake=1, tekka=1, toro=1)^-4 *
#> (anago=1, ebi=1, kappa=1, maguro=1, sake=1, tekka=1, toro=1, uni=1)^-8
#> * (anago=1, ebi=1, kappa=1, maguro=1, tamago=1, tekka=1, toro=1)^-1 *
#> (anago=1, ebi=1, kappa=1, maguro=1, tamago=1, tekka=1, toro=1,
#> uni=1)^-1 * (anago=1, ebi=1, kappa=1, maguro=1, tekka=1, toro=1)^-1 *
#> (anago=1, ebi=1, kappa=1, sake=1, tamago=1, tekka=1, toro=1, uni=1)^-1
#> * (anago=1, ebi=1, kappa=1, sake=1, tekka=1, toro=1)^-1 * (anago=1,
#> ebi=1, kappa=1, sake=1, tekka=1, toro=1, uni=1)^-3 * (anago=1, ebi=1,
#> kappa=1, tamago=1, tekka=1, toro=1)^-1 * (anago=1, ebi=1, kappa=1,
#> tekka=1, toro=1)^-1 * (anago=1, ebi=1, maguro=1, sake=1, tamago=1,
#> tekka=1, toro=1)^-2 * (anago=1, ebi=1, maguro=1, sake=1, tamago=1,
#> tekka=1, toro=1, uni=1)^-3 * (anago=1, ebi=1, maguro=1, sake=1,
#> tekka=1, toro=1)^-3 * (anago=1, ebi=1, maguro=1, sake=1, tekka=1,
#> toro=1, uni=1)^-2 * (anago=1, ebi=1, sake=1, tekka=1, toro=1)^-2 *
#> (anago=1, ika=1, kappa=1, maguro=1, sake=1, tamago=1, tekka=1, toro=1,
#> uni=1)^-4 * (anago=1, ika=1, kappa=1, maguro=1, sake=1, tekka=1,
#> toro=1)^-1 * (anago=1, ika=1, kappa=1, maguro=1, sake=1, tekka=1,
#> toro=1, uni=1)^-4 * (anago=1, ika=1, kappa=1, maguro=1, tamago=1,
#> tekka=1, toro=1, uni=1)^-2 * (anago=1, ika=1, kappa=1, sake=1, tekka=1,
#> toro=1, uni=1)^-1 * (anago=1, ika=1, kappa=1, tamago=1, tekka=1,
#> toro=1)^-1 * (anago=1, ika=1, maguro=1, sake=1, tekka=1, toro=1,
#> uni=1)^-1 * (anago=1, ika=1, sake=1, tekka=1, toro=1, uni=1)^-1 *
#> (anago=1, kappa=1, maguro=1, sake=1, tamago=1, tekka=1, toro=1,
#> uni=1)^-3 * (anago=1, kappa=1, maguro=1, sake=1, tekka=1, toro=1,
#> uni=1)^-2 * (anago=1, kappa=1, sake=1, tamago=1, tekka=1, toro=1,
#> uni=1)^-1 * (anago=1, kappa=1, tekka=1, toro=1)^-1 * (anago=1,
#> maguro=1, sake=1, tekka=1, toro=1, uni=1)^-1 * (anago=1, sake=1,
#> tamago=1, tekka=1, toro=1, uni=1)^-1 * (anago=1, tekka=1, toro=1)^-1 *
#> (anago=2.451008)^14 * (anago=2.451008, ebi=1, ika=1, kappa=1, maguro=1,
#> sake=1, tamago=1, tekka=2.451008, uni=1)^-2 * (anago=2.451008, ebi=1,
#> ika=1, kappa=1, maguro=1, sake=1, tamago=1, toro=2.451008, uni=1)^-1 *
#> (anago=2.451008, ebi=1, ika=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, uni=1)^-2 * (anago=2.451008, ebi=1, ika=1, kappa=1,
#> sake=1, tamago=1, tekka=2.451008, uni=1)^-2 * (anago=2.451008, ebi=1,
#> ika=1, kappa=1, sake=1, tekka=2.451008, uni=1)^-1 * (anago=2.451008,
#> ebi=1, ika=1, kappa=1, tekka=2.451008)^-1 * (anago=2.451008, ebi=1,
#> ika=1, kappa=1, tekka=2.451008, uni=1)^-1 * (anago=2.451008, ebi=1,
#> ika=1, maguro=1, sake=1, tamago=1, toro=2.451008, uni=1)^-1 *
#> (anago=2.451008, ebi=1, ika=1, maguro=1, tamago=1, toro=2.451008,
#> uni=1)^-1 * (anago=2.451008, ebi=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008)^-1 * (anago=2.451008, ebi=1, kappa=1, sake=1,
#> tekka=2.451008)^-1 * (anago=2.451008, ebi=1, maguro=1, sake=1,
#> tekka=2.451008)^-1 * (anago=2.451008, ebi=1, sake=1)^-1 *
#> (anago=2.451008, ebi=1, sake=1, tekka=2.451008)^-1 * (anago=2.451008,
#> ika=1, kappa=1, maguro=1, sake=1, tamago=1, tekka=2.451008)^-1 *
#> (anago=2.451008, ika=1, kappa=1, maguro=1, sake=1, tamago=1,
#> tekka=2.451008, uni=1)^-1 * (anago=2.451008, ika=1, kappa=1, maguro=1,
#> tamago=1, toro=2.451008, uni=1)^-1 * (anago=2.451008, ika=1, kappa=1,
#> tamago=1, tekka=2.451008)^-1 * (anago=2.451008, ika=1, kappa=1,
#> tamago=1, toro=2.451008, uni=1)^-1 * (anago=2.451008, ika=1,
#> tamago=1)^-1 * (anago=2.451008, ika=1, tamago=1, toro=2.451008,
#> uni=1)^-1 * (anago=2.451008, ika=1, tamago=1, uni=1)^-1 *
#> (anago=2.451008, kappa=1)^-1 * (anago=2.451008, kappa=1, maguro=1,
#> sake=1, tamago=1)^-1 * (anago=2.451008, kappa=1, maguro=1, sake=1,
#> tamago=1, tekka=2.451008)^-1 * (anago=2.451008, kappa=1, maguro=1,
#> sake=1, tekka=2.451008, uni=1)^-1 * (anago=2.451008, kappa=1, maguro=1,
#> tekka=2.451008)^-1 * (anago=2.451008, kappa=1, maguro=1,
#> tekka=2.451008, uni=1)^-1 * (anago=2.451008, kappa=1, sake=1,
#> tekka=2.451008)^-1 * (anago=2.451008, kappa=1, tamago=1,
#> tekka=2.451008)^-1 * (anago=2.451008, kappa=1, tekka=2.451008)^-2 *
#> (anago=2.451008, maguro=1, sake=1, tekka=2.451008, uni=1)^-1 *
#> (ebi=1)^43 * (ebi=1, ika=1)^-2 * (ebi=1, ika=1, kappa=1, maguro=1,
#> sake=1, tamago=1, tekka=2.451008, toro=2.451008)^-2 * (ebi=1, ika=1,
#> kappa=1, maguro=1, sake=1, tamago=1, tekka=2.451008, toro=2.451008,
#> uni=1)^-2 * (ebi=1, ika=1, kappa=1, maguro=1, sake=1, tamago=1,
#> tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008)^-2 * (ebi=1, ika=1, kappa=1, maguro=1,
#> sake=1, tekka=2.451008, toro=2.451008, uni=1)^-7 * (ebi=1, ika=1,
#> kappa=1, maguro=1, sake=1, tekka=2.451008, uni=1)^-5 * (ebi=1, ika=1,
#> kappa=1, maguro=1, sake=1, uni=1)^-1 * (ebi=1, ika=1, kappa=1,
#> maguro=1, tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, kappa=1, sake=1,
#> tamago=1, tekka=2.451008, uni=1)^-2 * (ebi=1, ika=1, kappa=1, sake=1,
#> tekka=2.451008)^-1 * (ebi=1, ika=1, kappa=1, sake=1, tekka=2.451008,
#> toro=2.451008)^-2 * (ebi=1, ika=1, kappa=1, sake=1, tekka=2.451008,
#> toro=2.451008, uni=1)^-1 * (ebi=1, ika=1, kappa=1, sake=1,
#> tekka=2.451008, uni=1)^-3 * (ebi=1, ika=1, kappa=1, tekka=2.451008)^-2
#> * (ebi=1, ika=1, kappa=1, tekka=2.451008, toro=2.451008)^-2 * (ebi=1,
#> ika=1, kappa=1, tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, maguro=1,
#> sake=1, tamago=1, tekka=2.451008)^-1 * (ebi=1, ika=1, maguro=1, sake=1,
#> tamago=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1, ika=1, maguro=1,
#> sake=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1, ika=1, maguro=1,
#> sake=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (ebi=1, ika=1,
#> maguro=1, sake=1, tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, maguro=1,
#> tamago=1, toro=2.451008, uni=1)^-1 * (ebi=1, ika=1, maguro=1,
#> tekka=2.451008)^-1 * (ebi=1, ika=1, maguro=1, tekka=2.451008,
#> toro=2.451008)^-1 * (ebi=1, ika=1, sake=1)^-1 * (ebi=1, ika=1, sake=1,
#> tamago=1)^-1 * (ebi=1, ika=1, sake=1, tamago=1, tekka=2.451008)^-1 *
#> (ebi=1, ika=1, sake=1, tekka=2.451008)^-1 * (ebi=1, ika=1, sake=1,
#> tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, tamago=1, toro=2.451008,
#> uni=1)^-1 * (ebi=1, ika=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1,
#> ika=1, tekka=2.451008, uni=1)^-1 * (ebi=1, ika=1, toro=2.451008)^-1 *
#> (ebi=1, ika=1, uni=1)^-1 * (ebi=1, kappa=1)^-1 * (ebi=1, kappa=1,
#> maguro=1)^-1 * (ebi=1, kappa=1, maguro=1, sake=1, tamago=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-1 * (ebi=1, kappa=1, maguro=1,
#> sake=1, tekka=2.451008)^-4 * (ebi=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008)^-5 * (ebi=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-9 * (ebi=1, kappa=1, maguro=1,
#> sake=1, tekka=2.451008, uni=1)^-4 * (ebi=1, kappa=1, maguro=1, sake=1,
#> uni=1)^-1 * (ebi=1, kappa=1, maguro=1, tekka=2.451008)^-1 * (ebi=1,
#> kappa=1, maguro=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1, kappa=1,
#> maguro=1, tekka=2.451008, uni=1)^-1 * (ebi=1, kappa=1, sake=1,
#> tekka=2.451008)^-4 * (ebi=1, kappa=1, sake=1, tekka=2.451008,
#> toro=2.451008)^-1 * (ebi=1, kappa=1, sake=1, tekka=2.451008,
#> toro=2.451008, uni=1)^-3 * (ebi=1, kappa=1, sake=1, tekka=2.451008,
#> uni=1)^-3 * (ebi=1, kappa=1, sake=1, uni=1)^-1 * (ebi=1, kappa=1,
#> tamago=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1, kappa=1,
#> tekka=2.451008)^-2 * (ebi=1, kappa=1, tekka=2.451008, uni=1)^-1 *
#> (ebi=1, kappa=1, uni=1)^-1 * (ebi=1, maguro=1)^-2 * (ebi=1, maguro=1,
#> sake=1)^-1 * (ebi=1, maguro=1, sake=1, tekka=2.451008)^-3 * (ebi=1,
#> maguro=1, sake=1, tekka=2.451008, toro=2.451008)^-4 * (ebi=1, maguro=1,
#> sake=1, tekka=2.451008, toro=2.451008, uni=1)^-4 * (ebi=1, maguro=1,
#> sake=1, tekka=2.451008, uni=1)^-1 * (ebi=1, maguro=1, tekka=2.451008,
#> toro=2.451008)^-2 * (ebi=1, maguro=1, tekka=2.451008, toro=2.451008,
#> uni=1)^-1 * (ebi=1, maguro=1, toro=2.451008)^-1 * (ebi=1, maguro=1,
#> toro=2.451008, uni=1)^-1 * (ebi=1, sake=1)^-1 * (ebi=1, sake=1,
#> tekka=2.451008)^-4 * (ebi=1, sake=1, tekka=2.451008, toro=2.451008)^-4
#> * (ebi=1, sake=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (ebi=1,
#> sake=1, tekka=2.451008, uni=1)^-1 * (ebi=1, sake=1, uni=1)^-1 * (ebi=1,
#> tamago=1, tekka=2.451008, toro=2.451008)^-1 * (ebi=1,
#> tekka=2.451008)^-3 * (ebi=1, toro=2.451008, uni=1)^-1 * (ebi=1,
#> uni=1)^-1 * (ika=1)^45 * (ika=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008)^-1 * (ika=1, kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-2 * (ika=1, kappa=1, maguro=1,
#> sake=1, tekka=2.451008, uni=1)^-2 * (ika=1, kappa=1, maguro=1,
#> tamago=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (ika=1, kappa=1,
#> sake=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (ika=1, kappa=1,
#> tekka=2.451008)^-1 * (ika=1, maguro=1, sake=1, tekka=2.451008)^-2 *
#> (ika=1, maguro=1, sake=1, tekka=2.451008, toro=2.451008)^-1 * (ika=1,
#> maguro=1, sake=1, tekka=2.451008, uni=1)^-1 * (ika=1, maguro=1,
#> tamago=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (ika=1, maguro=1,
#> tekka=2.451008)^-2 * (ika=1, maguro=1, tekka=2.451008, toro=2.451008,
#> uni=1)^-1 * (ika=1, sake=1, tekka=2.451008)^-2 * (ika=1, sake=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-2 * (ika=1, sake=1,
#> tekka=2.451008, uni=1)^-2 * (ika=1, tamago=1)^-1 * (ika=1, tamago=1,
#> toro=2.451008, uni=1)^-1 * (ika=1, tekka=2.451008)^-3 * (ika=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-2 * (ika=1, tekka=2.451008,
#> uni=1)^-3 * (ika=1, toro=2.451008)^-1 * (ika=1, toro=2.451008,
#> uni=1)^-1 * (ika=1, uni=1)^-4 * (kappa=1)^46 * (kappa=1, maguro=1)^-1 *
#> (kappa=1, maguro=1, sake=1, tamago=1)^-1 * (kappa=1, maguro=1, sake=1,
#> tamago=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (kappa=1,
#> maguro=1, sake=1, tekka=2.451008)^-3 * (kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008)^-2 * (kappa=1, maguro=1, sake=1,
#> tekka=2.451008, toro=2.451008, uni=1)^-2 * (kappa=1, maguro=1, sake=1,
#> tekka=2.451008, uni=1)^-3 * (kappa=1, maguro=1, tekka=2.451008)^-2 *
#> (kappa=1, sake=1, tekka=2.451008)^-3 * (kappa=1, tekka=2.451008)^-3 *
#> (kappa=1, tekka=2.451008, uni=1)^-1 * (kappa=1, uni=1)^-1 *
#> (maguro=1)^47 * (maguro=1, sake=1)^-2 * (maguro=1, sake=1, tamago=1)^-1
#> * (maguro=1, sake=1, tamago=1, tekka=2.451008, toro=2.451008, uni=1)^-1
#> * (maguro=1, sake=1, tamago=1, tekka=2.451008, uni=1)^-1 * (maguro=1,
#> sake=1, tekka=2.451008)^-3 * (maguro=1, sake=1, tekka=2.451008,
#> toro=2.451008)^-1 * (maguro=1, sake=1, tekka=2.451008, uni=1)^-3 *
#> (maguro=1, sake=1, uni=1)^-1 * (maguro=1, tamago=1)^-1 * (maguro=1,
#> tamago=1, tekka=2.451008, uni=1)^-1 * (maguro=1, tekka=2.451008)^-2 *
#> (maguro=1, tekka=2.451008, toro=2.451008)^-1 * (maguro=1,
#> tekka=2.451008, uni=1)^-1 * (maguro=1, uni=1)^-1 * (sake=1)^43 *
#> (sake=1, tamago=1, tekka=2.451008, toro=2.451008, uni=1)^-1 * (sake=1,
#> tekka=2.451008)^-10 * (sake=1, tekka=2.451008, toro=2.451008)^-4 *
#> (sake=1, tekka=2.451008, toro=2.451008, uni=1)^-2 * (sake=1,
#> tekka=2.451008, uni=1)^-2 * (sake=1, toro=2.451008)^-1 * (sake=1,
#> uni=1)^-2 * (tamago=1)^49 * (tamago=1, tekka=2.451008)^-1 * (tamago=1,
#> tekka=2.451008, toro=2.451008)^-1 * (tamago=1, tekka=2.451008,
#> uni=1)^-1 * (tekka=1)^2 * (tekka=2.451008)^34 * (tekka=2.451008,
#> toro=2.451008)^-3 * (tekka=2.451008, uni=1)^-2 * (toro=1)^12 *
#> (toro=2.451008)^35 * (uni=1)^44)
sushi_maxp
#>      anago        ebi        ika      kappa     maguro       sake     tamago 
#> 0.14563923 0.07050827 0.11348746 0.08814784 0.08609345 0.05453715 0.28260641 
#>      tekka       toro        uni 
#> 0.01716808 0.05726656 0.08454555