Results from the NOCS volleyball league

Results from the NOCS volleyball league. Object volleyball_table is a matrix in which each column corresponds to a player and each row corresponds to a volleyball set; volleyball is the corresponding likelihood function in the form of a hyper2 distribution.

data(volleyball)

Details

A volleyball set is a Bernoulli trial between two disjoint subsets of the players. The two subsets are denoted (after the game) as the “winners” and the “losers”: these are denoted by 1 and 0 respectively.

Thus the first line reads of volleyball_results reads:

 p1  p2  p3  p4  p5  p6  p7  p8  p9
 1    0  NA   1   0   0  NA   1  NA

showing that the teams were p1, p4 and p8 against p2, p5 and p6; players p3, p7 and p9 did not play.

These datasets illustrate the fact that such Bernoulli trials are only weakly informative.

These objects can be generated by running script inst/volleyball.Rmd, which includes some further discussion and technical documentation and creates file volleyball.rda which resides in the data/ directory.

The dataset is used in an example at zipf.Rd: the players' strengths are not Zipf.

Source

Volleyball games at NOCS, 2006-2008

References

Robin K. S. Hankin (2010). “A Generalization of the Dirichlet Distribution”, Journal of Statistical Software, 33(11), 1-18

Author

Robin K. S. Hankin

See also

Examples

volleyball == volley(volleyball_table)  # should be TRUE
#> [1] TRUE