Dataset from four Scottish football clubs

These objects refer to results from football matches among four Scottish clubs: Rangers, Celtic, Livinston, Falkirk. A detailed analysis is presented in inst/home_advantage.Rmd, which creates the objects documented here from scratch.

Object RCLF3_table is a table of results for the four clubs: home wins, draws, and away wins. RCLF3_lambda_max for the maximum likelihood estimate for \(\lambda\).

Object RCLF3 is a hyper3 likelihood function for the table of won games only [that is, excluding draws], using RCLF3_lambda_max for the value of \(\lambda\).

Object RCLF3_maxp is its evaluate at the MLE for lambda

data(RCLF)

Details

The objects are created by inst/home_advantage.Rmd.

Source

Data obtained from www.worldfootball.net

Author

Robin K. S. Hankin

Examples

data(RCLF)
RCLF3_table
#> $home_wins
#>             @away
#> @home        Rangers Celtic Falkirk Livingston
#>   Rangers         NA     81      69         15
#>   Celtic          76     NA      61         15
#>   Falkirk         17     17      NA          6
#>   Livingston       2      2       3         NA
#> 
#> $away_wins
#>             @away
#> @home        Rangers Celtic Falkirk Livingston
#>   Rangers         NA     42       2          0
#>   Celtic          46     NA       8          0
#>   Falkirk         57     43      NA          8
#>   Livingston      12     11       9         NA
#> 
#> $draws
#>             @away
#> @home        Rangers Celtic Falkirk Livingston
#>   Rangers         NA     44      14          2
#>   Celtic          47     NA      10          3
#>   Falkirk         12     19      NA          4
#>   Livingston       4      4       4         NA
#> 
RCLF3_maxp
#>     Celtic    Falkirk Livingston    Rangers 
#> 0.41160418 0.07357108 0.04249528 0.47232945 
RCLF3_lambda_max
#> [1] 1.773893

loglik(RCLF3_maxp,RCLF3)
#> [1] -305.933
equalp.test(RCLF3)
#> 
#> 	Constrained support maximization
#> 
#> data:  RCLF3
#> null hypothesis: Celtic = Falkirk = Livingston = Rangers
#> null estimate:
#>     Celtic    Falkirk Livingston    Rangers 
#>       0.25       0.25       0.25       0.25 
#> (argmax, constrained optimization)
#> Support for null:  -405.5553 + K
#> 
#> alternative hypothesis:  sum p_i=1 
#> alternative estimate:
#>     Celtic    Falkirk Livingston    Rangers 
#> 0.41160418 0.07357108 0.04249528 0.47232945 
#> (argmax, free optimization)
#> Support for alternative:  -305.933 + K
#> 
#> degrees of freedom: 3
#> support difference = 99.62227
#> p-value: 6.143175e-43 
#>