Random hyper2 objects

Random hyper2 loglikelihood functions, intended as quick “get you going” examples

rhyper2(n = 8, s = 5, pairs = TRUE, teams = TRUE, race = TRUE, pnames)

Arguments

n

Number of competitors, treated as even

s

Integer, Measure of the complexity of the log likelihood function

pairs,teams,race

Boolean, indicating whether or not to include different observations

pnames

Character vector of names, if missing interpret as letters; set to NA meaning no names

Author

Robin K. S. Hankin

Note

Function rhyper2() returns a likelihood function based on random observations. To return a random probability vector drawn from a from a given (normalized) likelihood function, use rp().

See also

Examples

rhyper2()
#> log(a^5 * (a + b)^-1 * (a + b + c + d + e + f + g)^-3 * (a + b + c + d
#> + e + f + g + h)^-10 * (a + b + c + e + f + g)^-1 * (a + b + c + e +
#> g)^-1 * (a + b + c + h) * (a + b + d + e + f + g)^-1 * (a + b + d +
#> f)^-1 * (a + b + d + f + g)^-1 * (a + b + f)^-1 * (a + c)^-1 * (a + c +
#> d + e + f)^-1 * (a + c + d + e + f + g)^-1 * (a + c + d + e + f + g +
#> h)^-1 * (a + c + d + f + g + h)^-1 * (a + c + d + g)^-1 * (a + c + d +
#> g + h)^-1 * (a + c + e)^-1 * (a + c + e + f)^-1 * (a + c + e + g)^-1 *
#> (a + c + g)^-1 * (a + d)^-1 * (a + f)^-1 * b^5 * (b + c + d + e) * (b +
#> c + d + e + f + g + h)^-1 * (b + c + d + f + g + h)^-1 * (b + d + f +
#> g) * (b + d + f + g + h)^-1 * (b + d + g + h)^-1 * c^4 * (c + d)^-1 *
#> (c + e)^-2 * (c + e + g)^-1 * (c + f)^-1 * (c + h)^-1 * d^6 * (d + f +
#> g + h)^2 * (d + g + h)^-1 * e^3 * f^7 * g^5 * (g + h)^-1 * h^5)
rp(2,icons)
#>             NB         L        PB       THC        OA      WAIS
#> [1,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [2,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667