Convert various datasets to support functions.

# S3 method for class 'ordertable'
suppfun(x)
# S3 method for class 'ranktable'
suppfun(x)
# S3 method for class 'preftable'
suppfun(x)
suppfun(x)

Arguments

x

Dataset to be converted to a support function

Details

Function suppfun() is intended as a consistent S3 generic approach to converting datasets to support functions. I am gradually going to switch out use of ordertable2supp() in favour of suppfun() in the docs.

If given a data frame, suppfun() will not guess whether it is to be interpreted as an ordertable or a ranktable, and return an error.

Value

Generally, return a support function

Author

Robin K. S. Hankin

Examples

suppfun(pentathlon_table)
#> log(Capalini^5 * (Capalini + Cerkovskis + Meliakh + Michalik + Moiseev
#> + Walther)^-1 * (Capalini + Cerkovskis + Meliakh + Michalik + Moiseev +
#> Walther + Zadneprovskis)^-5 * (Capalini + Cerkovskis + Meliakh +
#> Michalik + Walther)^-1 * (Capalini + Cerkovskis + Meliakh + Michalik +
#> Walther + Zadneprovskis)^-2 * (Capalini + Cerkovskis + Michalik +
#> Moiseev + Walther + Zadneprovskis)^-2 * (Capalini + Cerkovskis +
#> Moiseev + Walther + Zadneprovskis)^-2 * (Capalini + Meliakh)^-1 *
#> (Capalini + Meliakh + Michalik + Moiseev)^-1 * (Capalini + Meliakh +
#> Michalik + Moiseev + Walther)^-1 * (Capalini + Meliakh + Michalik +
#> Walther)^-1 * (Capalini + Meliakh + Walther)^-1 * (Capalini + Moiseev +
#> Walther + Zadneprovskis)^-1 * Cerkovskis^3 * (Cerkovskis + Meliakh +
#> Michalik + Walther + Zadneprovskis)^-1 * (Cerkovskis + Meliakh +
#> Michalik + Zadneprovskis)^-1 * (Cerkovskis + Michalik)^-1 * (Cerkovskis
#> + Michalik + Zadneprovskis)^-1 * (Cerkovskis + Moiseev)^-1 *
#> (Cerkovskis + Moiseev + Walther + Zadneprovskis)^-1 * (Cerkovskis +
#> Moiseev + Zadneprovskis)^-1 * Meliakh^4 * (Meliakh + Michalik)^-1 *
#> (Meliakh + Michalik + Moiseev)^-1 * Michalik^4 * Moiseev^5 * (Moiseev +
#> Walther + Zadneprovskis)^-1 * Walther^4 * (Walther + Zadneprovskis)^-1
#> * Zadneprovskis^5)