Match outcomes from repeated doubles tennis matches

Match outcomes from repeated doubles tennis matches

data(tennis)

Format

A hyper2 object corresponding to the match outcomes listed below.

Details

There are four players, \(p_1\) to \(p_4\). These players play doubles tennis matches with the following results:

match	score

\(\lbrace p_1,p_2\rbrace\) vs \(\lbrace p_3,p_4\rbrace\)	9-2

\(\lbrace p_1,p_3\rbrace\) vs \(\lbrace p_2,p_4\rbrace\)	4-4

\(\lbrace p_1,p_4\rbrace\) vs \(\lbrace p_2,p_3\rbrace\)	6-7

\(\lbrace p_1\rbrace\) vs \(\lbrace p_3\rbrace\)	10-14

\(\lbrace p_2\rbrace\) vs \(\lbrace p_3\rbrace\)	12-14

\(\lbrace p_1\rbrace\) vs \(\lbrace p_4\rbrace\)	10-14

\(\lbrace p_2\rbrace\) vs \(\lbrace p_4\rbrace\)	11-10

\(\lbrace p_3\rbrace\) vs \(\lbrace p_4\rbrace\)	13-13

It is suspected that \(p_1\) and \(p_2\) have some form of team cohesion and play better when paired than when either solo or with other players. As the scores show, each player and, apart from p1-p2, each doubles partnership, is of approximately the same strength.

Dataset tennis gives the appropriate likelihood function for the players' strengths; and dataset tennis_ghost gives the appropriate likelihood function if the extra strength due to team cohesion of \(\lbrace p_1,p_2\rbrace\) is represented by a ghost player.

These objects can be generated by running script inst/tennis.Rmd, which includes some further discussion and technical documentation and creates file tennis.rda which resides in the data/ directory.

Source

Doubles tennis matches at NOCS, Jan-May 2008

References

Robin K. S. Hankin (2010). “A Generalization of the Dirichlet Distribution”, Journal of Statistical Software, 33(11), 1-18

Examples

summary(tennis)
#> A hyper2 object of size 4.
#> pnames:  p1 p2 p3 p4 
#> Number of brackets: 11 
#> Sum of powers: 0 
#> 
#> Table of bracket lengths:
#> 1 2 4 
#> 4 6 1 
#> 
#> Table of powers:
#> -32 -24 -20 -19 -18 -17   9  20  23  37  41 
#>   1   1   1   1   1   1   1   1   1   1   1 

tennis |> psubs(c("Federer","Laver","Graf","Navratilova"))
#> log(Federer^20 * (Federer + Graf)^-20 * (Federer + Graf + Laver +
#> Navratilova)^-32 * (Federer + Laver)^9 * (Federer + Navratilova)^-18 *
#> Graf^41 * (Graf + Laver)^-19 * (Graf + Navratilova)^-24 * Laver^23 *
#> (Laver + Navratilova)^-17 * Navratilova^37)

## Following line commented out because it takes too long:
# specificp.gt.test(tennis_ghost,"G",0)